+++ system/bta/dm/bta_dm_act.cc
@@ -125,7 +125,7 @@ namespace {
 
 struct WaitForAllAclConnectionsToDrain {
   uint64_t time_to_wait_in_ms;
-  unsigned long TimeToWaitInMs() const { return static_cast<unsigned long>(time_to_wait_in_ms); }
+  uint64_t TimeToWaitInMs() const { return time_to_wait_in_ms; }
   void* AlarmCallbackData() const { return const_cast<void*>(static_cast<const void*>(this)); }
 
   static const WaitForAllAclConnectionsToDrain* FromAlarmCallbackData(void* data);
@@ -339,8 +339,7 @@ void bta_dm_disable() {
         bta_dm_disable_conn_down_timer_cback(nullptr);
         break;
       default:
-        log::debug("Set timer to delay disable initiation:{} ms",
-                   static_cast<unsigned long>(disable_delay_ms));
+        log::debug("Set timer to delay disable initiation:{} ms", disable_delay_ms);
         alarm_set_on_mloop(bta_dm_cb.disable_timer, disable_delay_ms,
                            bta_dm_disable_conn_down_timer_cback, nullptr);
     }
